Extract NotificationOutputExtractor logic
