+++ examples/templates/output/channel-result.twig
