+++ examples/using-all-channels/app.php
@@ -59,5 +59,5 @@ $result = $notifier->notify(NotifierInput::new(
 
 (new NotifierOutput(
     new EchoOutput(),
-    new Environment(new FilesystemLoader('templates'))
+    new Environment(new FilesystemLoader('../templates'))
 ))->write($result);
