+++ examples/using-email-channel/app.php
@@ -50,5 +50,5 @@ $result = $notifier->notify(NotifierInput::new(
 
 (new NotifierOutput(
     new EchoOutput(),
-    new Environment(new FilesystemLoader('templates'))
+    new Environment(new FilesystemLoader('../templates'))
 ))->write($result);
