+++ examples/using-slack-channel/app.php
@@ -46,5 +46,5 @@ $result = $notifier->notify(NotifierInput::new(
 
 (new NotifierOutput(
     new EchoOutput(),
-    new Environment(new FilesystemLoader('templates'))
+    new Environment(new FilesystemLoader('../templates'))
 ))->write($result);
