+++ src/ScrumMaster/IO/NotifierOutput.php
@@ -5,8 +5,6 @@ declare(strict_types=1);
 namespace Chemaclass\ScrumMaster\IO;
 
 use Chemaclass\ScrumMaster\Channel\ChannelResult;
-use Chemaclass\ScrumMaster\Channel\ReadModel\ChannelIssue;
-use Chemaclass\ScrumMaster\Common\Request;
 use Twig\Environment;
 
 final class NotifierOutput
@@ -33,60 +31,16 @@ final class NotifierOutput
 
     private function writeChannel(string $channelName, ChannelResult $result): void
     {
-        $notificationTitles = $this->buildNotificationTitles($result);
-        $notificationSuccessful = $this->buildNotificationSuccessful($result);
-        $notificationFailed = $this->buildNotificationFailed($result);
+        $outputExtractor = new NotificationOutputExtractor($result);
 
-        $render = $this->twig->render('ticket_status.twig', [
+        $render = $this->twig->render('output/channel-result.twig', [
             'channelName' => $channelName,
             'result' => $result,
-            'notificationTitles' => $notificationTitles,
-            'notificationSucessful' => $notificationSuccessful,
-            'notificationFailed' => $notificationFailed,
+            'notificationTitles' => $outputExtractor->titles(),
+            'notificationSuccessful' => $outputExtractor->successful(),
+            'notificationFailed' => $outputExtractor->failed(),
         ]);
 
-        $lines = explode("\n", $render);
-
-        foreach ($lines as $line) {
-            $this->output->writeln($line);
-        }
-    }
-
-    private function buildNotificationTitles(ChannelResult $result): string
-    {
-        $notificationTitles = [];
-
-        /** @var ChannelIssue $channelIssue */
-        foreach ($result->channelIssues() as $statusCode => $channelIssue) {
-            $notificationTitles[] = (null !== $channelIssue->displayName())
-                ? "$statusCode: {$channelIssue->displayName()}"
-                : $statusCode;
-        }
-
-        return implode(', ', $notificationTitles);
-    }
-
-    private function buildNotificationSuccessful(ChannelResult $result): string
-    {
-        $notificationSuccessful = array_keys(array_filter(
-            $result->channelIssues(),
-            function (ChannelIssue $slackTicket) {
-                return Request::HTTP_OK === $slackTicket->responseStatusCode();
-            }
-        ));
-
-        return implode(', ', $notificationSuccessful);
-    }
-
-    private function buildNotificationFailed(ChannelResult $result): string
-    {
-        $notificationFailed = array_keys(array_filter(
-            $result->channelIssues(),
-            function (ChannelIssue $slackTicket) {
-                return Request::HTTP_OK !== $slackTicket->responseStatusCode();
-            }
-        ));
-
-        return implode(', ', $notificationFailed);
+        $this->output->writeln($render);
     }
 }
